Added id stuff and some cleanup on target lists
Also I have changed every moment where a structure from logData was passed directly to the "data" property of the component. The reason is that the "data" property and "props" property of a component are reactive which means Vue puts dynamic gets/sets on every attribute of the object. The logData object is never modified once it's given to the main component but it's read very often, having dynamic gets/sets (+Vue listening to changes on it) can negatively impact other components reading players or mechanics a lot. The elements returned in the "computed" attribute's methods are not reactive however and a good way to communicate logData stuff into the template when applicable.
That's why the object "simpleLogData" exists in ei.js, it's in order to make only a very small and necessary subset of logData editable/reactive.