Split the executalbe into two projects, windowed (old) and cli. Moved Settings into PArserCommon to share it between cli and windowed. Changed the test dependencies to ParserCommon wich allows the tests to also target non windows frameworks.
Updated the language version of subprojects to 10 for upcoming changes.
Had to change the .editorconfig, because var declaraion was in violation of an arror rule. Probably a change in what is considered for the specific rule between versions.

I really tried to make this work as one codebase, but VS _really_ doesnt wnat you to build two different executables form the same codebase.
I again almost got it working, but in the end the Forms designer wouldnt load.
If you ever wnat to attempt the unification you need to
 - change it to a target framework set instead of a specific one
 - set different output paths
 - set different intermediate paths before the actual project file gets laoded. This requires adding a Build.Directory.props (iirc the name)
 - splitting project packages and includes based on target framework (`<Compile Remove="" />` forms and resources in the generic version)
 - find a way to get the forms designer working